// TicketingSystem/ticketing-system/src/app/api/models/player-header.ts
/* tslint:disable */
/* eslint-disable */
export interface PlayerHeader {
  birthDate?: Date;
  firstName?: null | string;
  heigth?: number;
  id?: number;
  jerseyNumber?: number;
  lastName?: null | string;
  nationality?: null | string;
  position?: null | string;
  source?: null | string;
  weigth?: number;
}